scheduler_trx: use stored block fn instead of calculating it.
When the block is passed up into the higher layers it it marked with the
frame number that points to the beginning of the block. At the moment
some strange calculations are used to calculate the beginning of the
block fn from the current fn that points at the end of the block:

PDTCH: fn != (fn + GSM_HYPERFRAME - 3) % GSM_HYPERFRAME
FACCH/F: fn != (fn + GSM_HYPERFRAME - 7) % GSM_HYPERFRAME
TCH/F: fn != (fn + GSM_HYPERFRAME - 7) % GSM_HYPERFRAME

Despite the fact that those equations do not look very trustworthy, it
is also not necessary to calculate the fn since the code stores the fn
of each beginning of a new block in chan_state->ul_first_fn (*first_fn),
so we can just use the stored fn when passing the block up. This also
makes the code more logical since the measurement indications already
use the stored frame number as well.

Change-Id: Ia27254bbf6e36426f7890ece6154dcd395673f63
Related: OS#2977