OML: Ensure minimum required attributes are set before accepting OPSTART
There are some vital configuration bits such as ARFCN, BSIC, channel
config, .. for which there are no reasonable defaults.  As a result,
the BSC must set those attributes before issuing OPSTART.

Prior to this patch we would blindly accept OPSTART and then transmit
on ARFCN 0, which is definitely not the intended behavior.

Closes: OS#3789
Change-Id: I3a818f8eceb6abef1b20d2b3892a749dbc9e4b05